refactor: restructure WeChat client module by splitting token, followers, and follower sync logic into reusable components
